fix(gateway): move session_manager tests to correct path
- Relocate session_manager_tests.rs to src/gateway/session_manager/tests.rs
- Fix E0583 error caused by incorrect module file location
- Remove duplicate #[cfg(test)] attribute

Source: CI
Type: fix
