+++ src/gateway/session_manager.rs
@@ -237,5 +237,4 @@ impl SessionManager {
 // ---------------------------------------------------------------------------
 
 #[cfg(test)]
-#[cfg(test)]
 mod tests;
